+++ src/Provider/DataSet/DeferredDataSetProvider.php
@@ -4,8 +4,8 @@ namespace webignition\BasilParser\Provider\DataSet;
 
 use webignition\BasilParser\Exception\NonRetrievableDataProviderException;
 use webignition\BasilParser\Exception\UnknownDataProviderException;
+use webignition\BasilParser\Exception\YamlLoaderException;
 use webignition\BasilParser\Loader\DataSetLoader;
-use webignition\BasilParser\Loader\YamlLoaderException;
 use webignition\BasilParser\Model\DataSet\DataSetInterface;
 
 class DeferredDataSetProvider implements DataSetProviderInterface
